[pass] Pass after all optimizations

+++ pass/log.c
@@ -7,8 +7,8 @@ void bbStartLogger(char* FileName, size_t *BbAddr) {
     fclose(DumpFile);
 }
 
-void binOpLogger(char* FileName, size_t *BbAddr, size_t *InstrAddr, int Result) {
+void binOpLogger(char* FileName, size_t *BbAddr, size_t *InstrAddr, long long int Result) {
     FILE *DumpFile = fopen(FileName, "a");
-    fprintf(DumpFile, "binop %p %p %d\n", BbAddr, InstrAddr, Result);
+    fprintf(DumpFile, "binop %p %p %lld\n", BbAddr, InstrAddr, Result);
     fclose(DumpFile);
 }
